requirements.txt

google-cloud-texttospeech
PyPDF2
textract

